fixed windows problems

+++ examples/kinect/data_reader.cxx
@@ -52,18 +52,18 @@ std::vector<point_3d> data_reader::get()
     std::vector<point_3d> points;
     points.reserve(img.cols * img.rows);
 
-    const double fx = 535.4;
-    const double fy = 539.2;
-    const double cx = 320.1;
-    const double cy = 247.6;
-    const double factor = 5000;
+    const float fx = 535.4f;
+    const float fy = 539.2f;
+    const float cx = 320.1f;
+    const float cy = 247.6f;
+    const float factor = 5000;
 
     for(int v = 0; v < img.rows; v++){
         for(int u = 0; u < img.cols; u++){
-            double val = img.at<uint16_t>(v,u);
-            double Z = val/factor;
-            double X = (u - cx) * Z / fx;
-            double Y = (v - cy) * Z / fy;
+            float val = img.at<uint16_t>(v,u);
+            float Z = val/factor;
+            float X = (u - cx) * Z / fx;
+            float Y = (v - cy) * Z / fy;
             if(u%8 == 0 && v%8 == 0 && Z != 0)
                 points.push_back(point_3d(X,Y,Z,u,v));
         }
